+++ assets/javascripts/kitten/components/atoms/icon-badge/index.js
@@ -16,8 +16,7 @@ const StyledBadge = styled.span`
   min-width: ${pxToRem(30)};
   min-height: ${pxToRem(30)};
   background-color: ${COLORS.primary1};
-
-  border-radius: var(--border-radius);
+  border-radius: ${pxToRem(30)};
 
   &.k-IconBadge--empty {
     border: ${pxToRem(2)} solid ${COLORS.line1};
@@ -81,7 +80,7 @@ const StyledBadge = styled.span`
   }
 
   &.k-IconBadge--hasBorderStyles {
-    border-width: var(--border-width);
+    border-width: var(--border-width, 0);
     border-color: var(--border-color);
     border-style: var(--border-style);
   }
@@ -96,10 +95,7 @@ export const IconBadge = ({
   big,
   huge,
   size,
-  borderRadius,
-  borderWidth,
-  borderColor,
-  borderStyle,
+  border,
   ...others
 }) => (
   <StyledBadge
@@ -109,17 +105,11 @@ export const IconBadge = ({
       'k-IconBadge--empty': empty,
       'k-IconBadge--big': big,
       'k-IconBadge--huge': huge,
-      'k-IconBadge--hasBorderStyles': 
-        [
-          borderRadius > 0,
-          borderWidth > 0,
-        ],
     })}
     style={{
-      '--border-radius': pxToRem(borderRadius),
-      '--border-width': pxToRem(borderWidth),
-      '--border-color': borderColor,
-      '--border-style': borderStyle,
+      '--border-width': pxToRem(border?.width),
+      '--border-style': border?.style,
+      '--border-color': border?.color,
     }}
     {...others}
   >
@@ -132,7 +122,6 @@ IconBadge.defaultProps = {
   valid: false,
   empty: false,
   size: 'normal',
-  borderRadius: 30,
   borderWidth: null,
   borderColor: null,
   borderStyle: null,
@@ -145,7 +134,6 @@ IconBadge.propTypes = {
   big: deprecated(PropTypes.bool, 'Use `size` prop instead.'),
   huge: deprecated(PropTypes.bool, 'Use `size` prop instead.'),
   size: PropTypes.oneOf(['tiny', 'normal', 'big', 'huge']),
-  borderRadius: PropTypes.number,
   borderWidth: PropTypes.number,
   borderColor: PropTypes.string,
   borderStyle: PropTypes.string,