+++ assets/javascripts/kitten/components/atoms/icon-badge/stories.js
@@ -31,22 +31,9 @@ export default {
       name: 'empty',
       control: 'boolean',
     },
-    borderRadius: {
-      name: 'borderRadius',
+    border: {
+      name: 'border',
       control: 'number',
-    },
-    borderWidth: {
-      name: 'borderWidth',
-      control: 'number',
-    },
-    borderColor: {
-      name: 'borderColor',
-      control: 'color',
-    },
-    borderStyle: {
-      name: 'borderStyle',
-      control: 'select',
-      options: ['solid', 'dotted', 'dashed'],
     },
     size: {
       name: 'size',
@@ -61,10 +48,6 @@ const args = {
   valid: false,
   empty: false,
   size: 'normal',
-  borderRadius: 30,
-  borderColor: null,
-  borderWidth: null,
-  borderStyle: null,
   children: <StarIcon color={COLORS.background1} />,
 }
 
@@ -76,8 +59,5 @@ export const WithBorder = args => <IconBadge {...args} />
 
 WithBorder.args = {
   ...args,
-  borderColor: '#222',
-  borderWidth: 2,
-  borderStyle: 'solid',
   size: 'tiny',
 }